backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
pgvector
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
email-validator
python-multipart
xlrd
pytest
pytest-asyncio
httpx
factory-boy
aiosqlite
